+++ src/main/java/io/bitbucket/plt/autotutor/racket/test/Plus.java
@@ -12,11 +12,11 @@ public class Plus extends Expression{
 
     @Override
     public String evaluate(List<Expression> list) {
-        int result = 0;
+        float result = 0;
         for (Expression e : list) {
-            result += Integer.valueOf(e.evaluate(this));
+            result += Float.valueOf(e.evaluate(this));
         }
-        return Integer.toString(result);
+        return Float.toString(result);
     }
 
     @Override
